+++ src/main/java/step4/domain/Cars.java
@@ -1,4 +1,7 @@
-package step4;
+package step4.domain;
+
+import step4.strategy.MoveStrategy;
+import step4.service.dto.Winners;
 
 import java.util.List;
 import java.util.NoSuchElementException;
