+++ src/main/java/step4/domain/Name.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.domain;
 
 import java.util.Objects;
