+++ src/main/java/step4/main.java
@@ -1,5 +1,14 @@
 package step4;
 
+import step4.service.Game;
+import step4.service.dto.GameInformation;
+import step4.strategy.MoveStrategy;
+import step4.strategy.NumberGenerator;
+import step4.strategy.RandomMoveStrategy;
+import step4.strategy.RandomNumberGenerator;
+import step4.view.InputView;
+import step4.view.ResultView;
+
 public class main {
     public static void main(String[] args) {
         InputView inputView = new InputView();
