+++ src/main/java/step4/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.strategy;
 
 public interface MoveStrategy {
