+++ src/main/java/step4/strategy/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.strategy;
 
 import java.util.Random;
