+++ src/main/java/step4/view/ResultView.java
@@ -1,4 +1,7 @@
-package step4;
+package step4.view;
+
+import step4.service.dto.GameHistory;
+import step4.service.dto.GameInformation;
 
 public class ResultView {
