+++ src/test/java/step4/GameHistoryTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step4.domain.*;
+import step4.service.dto.GameHistory;
+import step4.strategy.MoveStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
