+++ src/test/java/step4/GameInformationTest.java
@@ -3,6 +3,8 @@ package step4;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step4.service.dto.GameInformation;
+import step4.view.UserInput;
 
 import static org.assertj.core.api.Assertions.assertThat;
