+++ src/test/java/step4/NameTest.java
@@ -3,6 +3,7 @@ package step4;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step4.domain.Name;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
