+++ src/test/java/step4/NamesTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step4.domain.Name;
+import step4.domain.Names;
 
 import java.util.Arrays;
 import java.util.List;
