+++ src/test/java/step4/UserInputTest.java
@@ -2,8 +2,8 @@ package step4;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import step4.view.UserInput;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
